+++ dimas/src/com/liveliness_subscriber.rs
@@ -4,7 +4,7 @@
 //! A `LivelinessSubscriber` can optional subscribe on a delete message.
 
 // region:		--- modules
-use crate::{agent::TaskSignal, prelude::*};
+use crate::{prelude::*, utils::TaskSignal};
 use std::{
 	sync::{mpsc::Sender, Mutex},
 	time::Duration,
@@ -224,7 +224,8 @@ where
 		let c = self.storage.storage.clone();
 		let s = self.build()?;
 
-		let r = c.write()
+		let r = c
+			.write()
 			.map_err(|_| DimasError::ShouldNotHappen)?
 			.insert(s.key_expr.clone(), s);
 		Ok(r)
@@ -291,7 +292,7 @@ where
 		let p_cb = self.put_callback.clone();
 		let ctx = self.context.clone();
 		let key_expr = self.key_expr.clone();
-		tokio::spawn(async move {
+		tokio::task::spawn(async move {
 			if let Err(error) = run_initial(key_expr, p_cb, ctx).await {
 				error!("spawning initial liveliness failed with {error}");
 			};
@@ -303,22 +304,23 @@ where
 		let ctx = self.context.clone();
 		let key_expr = self.key_expr.clone();
 
-		self.handle.replace(tokio::spawn(async move {
-			#[cfg(feature = "liveliness")]
-			let key = key_expr.clone();
-			std::panic::set_hook(Box::new(move |reason| {
-				error!("liveliness subscriber panic: {}", reason);
+		self.handle
+			.replace(tokio::task::spawn(async move {
 				#[cfg(feature = "liveliness")]
-				if let Err(reason) = tx.send(TaskSignal::RestartLiveliness(key.clone())) {
-					error!("could not restart liveliness subscriber: {}", reason);
-				} else {
-					info!("restarting liveliness subscriber!");
+				let key = key_expr.clone();
+				std::panic::set_hook(Box::new(move |reason| {
+					error!("liveliness subscriber panic: {}", reason);
+					#[cfg(feature = "liveliness")]
+					if let Err(reason) = tx.send(TaskSignal::RestartLiveliness(key.clone())) {
+						error!("could not restart liveliness subscriber: {}", reason);
+					} else {
+						info!("restarting liveliness subscriber!");
+					};
+				}));
+				if let Err(error) = run_liveliness(key_expr, p_cb, d_cb, ctx).await {
+					error!("spawning liveliness subscriber failed with {error}");
 				};
 			}));
-			if let Err(error) = run_liveliness(key_expr, p_cb, d_cb, ctx).await {
-				error!("spawning liveliness subscriber failed with {error}");
-			};
-		}));
 	}
 
 	/// Stop a running LivelinessSubscriber