+++ dimas/src/prelude.rs
@@ -1,6 +1,5 @@
 // Copyright © 2023 Stephan Kunz
 
-
 //! Public interface of dimas. Typically it is sufficient to include the prelude with
 //! ```use dimas::prelude::*;```