+++ examples/liveliness/src/liveliness.rs
@@ -29,7 +29,7 @@ fn delete_subscription(ctx: &ArcContext<AgentProps>, id: &str) -> Result<()> {
 	Ok(())
 }
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();