+++ examples/pingpong/src/pong.rs
@@ -36,7 +36,7 @@ fn ping_received(ctx: &ArcContext<AgentProps>, message: Message) -> Result<()> {
 	Ok(())
 }
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();
@@ -48,12 +48,12 @@ async fn main() -> Result<()> {
 	let mut agent = Agent::new_with_prefix(Config::default(), properties, "examples")?;
 
 	// create publisher for topic "ping"
-	agent.publisher().msg_type("pong").add()?;
+	agent.publisher().topic("pong").add()?;
 
 	// listen for 'ping' messages
 	agent
 		.subscriber()
-		.msg_type("ping")
+		.topic("ping")
 		.put_callback(ping_received)
 		.add()?;