+++ examples/pubsub/src/subscriber.rs
@@ -24,7 +24,7 @@ fn hello_deletion(ctx: &ArcContext<AgentProps>) -> Result<()> {
 	Ok(())
 }
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();
@@ -38,7 +38,7 @@ async fn main() -> Result<()> {
 	// listen for 'hello' messages
 	agent
 		.subscriber()
-		.msg_type("hello")
+		.topic("hello")
 		.put_callback(hello_publishing)
 		.delete_callback(hello_deletion)
 		.add()?;