+++ examples/queries/src/query.rs
@@ -16,7 +16,7 @@ fn query_callback(_ctx: &ArcContext<AgentProps>, response: Response) -> Result<(
 	Ok(())
 }
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() -> Result<()> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();
@@ -30,7 +30,7 @@ async fn main() -> Result<()> {
 	// create publisher for topic "ping"
 	agent
 		.query()
-		.msg_type("query")
+		.topic("query")
 		.callback(query_callback)
 		.add()?;